EXC-614: Avoid race conditions in PUQ certification.
Prior to this PR, PUQ certification can run into race conditions. For example:

* `get_account` query call fetches a malicious account identifier.
* Before the result of the `get_account` update call is returned, the
  `ledger_sync_service` starts PUQing the ICP balance of the malicious
  account identifier.
* Response from the update call of `get_account` is returned, updating
  the UI to show the correct account identifier, but the balance
  remains that of the malicious account identifier.

This PR merges the `BalanceSyncService`, `TransactionSyncService`,
and `AccountSyncService` into one, and orchestrates the work such that
these race conditions never happen.
